Thêm tính năng - thêm tính năng 2 shop mua vé quay của npc Hải Đăng, chỉnh giới hạn vàng từ int sang long để tránh tràn số

+++ core/src/main/java/com/dang/dragonboy/du_lieu/DuLieuNguoiChoi.java
@@ -668,7 +668,7 @@ public class DuLieuNguoiChoi {
         this.GiamSatThuongNhanVat -= PtGiamSatThuong;
         this.GiamSatThuongNhanVat = Math.min(this.GiamSatThuongNhanVat,85);
     }
-    public void giamVang(int soLuong) {
+    public void giamVang(long soLuong) {
         this.vang -= soLuong;
     }
